+++ src/main/java/eu/gir/girsignals/EnumSignals.java
@@ -37,7 +37,7 @@ public class EnumSignals {
 
 	public enum ZS32 implements Offable<ZS32> {
 		OFF, A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, W, X, Y, Z, Z1, Z2, Z3, Z4, Z5, Z6, Z7,
-		Z8, Z9, Z10, Z11, Z12, Z13, Z14, Z15, Z16, Z17, Z18, Z19, Z20, ZS6, ZS8, ZS13;
+		Z8, Z9, Z10, Z11, Z12, Z13, Z14, Z15, Z16, Z17, Z18, Z19, Z20, ZS6, ZS8, ZS13, ZP8, ZP9;
 	}
 
 	public enum MAST_SIGN implements Offable<MAST_SIGN> {
@@ -69,7 +69,7 @@ public class EnumSignals {
 	public enum SH_LIGHT implements Offable<SH_LIGHT> {
 		OFF, SH0, SH1;
 	}
-	
+
 	public enum TRAMTYPE implements DefaultName<TRAMTYPE> {
 		TRAM, CAR, PEDESTRIAN
 	}
@@ -108,7 +108,11 @@ public class EnumSignals {
 	}
 
 	public enum RA implements DefaultName<RA> {
-		RA10, RA11A, RA11B, RA12;
+		RA10, RA11A, RA11B, RA12, RA6_9;
+	}
+
+	public enum RA_LIGHT implements Offable<RA_LIGHT> {
+		OFF, RA6, RA7, RA8, RA9;
 	}
 
 	public enum BUE implements DefaultName<BUE> {
@@ -139,6 +143,14 @@ public class EnumSignals {
 
 	public enum NE_ADDITION implements Offable<NE_ADDITION> {
 		OFF, PRE1, PRE2;
+	}
+
+	public enum WN_NORMAL implements Offable<WN_NORMAL> {
+		OFF, WN1, WN2, BLINK;
+	}
+
+	public enum WN_CROSS implements Offable<WN_CROSS> {
+		OFF, WN3, WN4, WN5, WN6, BLINK;
 	}
 
 }