+++ src/main/java/eu/gir/girsignals/blocks/boards/SignalRA.java
@@ -3,6 +3,7 @@ package eu.gir.girsignals.blocks.boards;
 import java.util.HashMap;
 
 import eu.gir.girsignals.EnumSignals.RA;
+import eu.gir.girsignals.EnumSignals.RA_LIGHT;
 import eu.gir.girsignals.SEProperty;
 import eu.gir.girsignals.SEProperty.ChangeableStage;
 import eu.gir.girsignals.blocks.Signal;
@@ -15,12 +16,16 @@ public class SignalRA extends Signal {
 	}
 
 	public static final SEProperty<RA> RATYPE = SEProperty.of("ratype", RA.RA10, ChangeableStage.GUISTAGE);
-	public static final SEProperty<Boolean> RALIGHT = SEProperty.of("ralight", false);
+	public static final SEProperty<Boolean> RALIGHT = SEProperty.of("ralight", false, ChangeableStage.APISTAGE,
+			t -> t.entrySet().stream().anyMatch((e -> e.getKey().equals(RATYPE) && e.getValue().equals(RA.RA11A))));
+	public static final SEProperty<RA_LIGHT> RALIGHTSIGNAL = SEProperty.of("ralightsignal", RA_LIGHT.OFF,
+			ChangeableStage.APISTAGE_NONE_CONFIG,
+			t -> t.entrySet().stream().anyMatch((e -> e.getKey().equals(RATYPE) && e.getValue().equals(RA.RA6_9))));
 
 	@Override
 	public int getHeight(final HashMap<SEProperty<?>, Object> map) {
 		final RA ra = (RA) map.get(RATYPE);
-		if(ra == null)
+		if (ra == null)
 			return super.getHeight(map);
 		switch (ra) {
 		case RA10:
@@ -28,9 +33,11 @@ public class SignalRA extends Signal {
 		case RA11A:
 		case RA11B:
 			return 3;
+		case RA6_9:
+			return 4;
 		default:
 			return super.getHeight(map);
 		}
 	}
-	
+
 }