+++ src/main/java/eu/gir/girsignals/guis/GuiSignalController.java
@@ -20,11 +20,14 @@ import static eu.gir.girsignals.guis.GuiPlacementtool.visible;
 
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Map.Entry;
 import java.util.function.Consumer;
 
 import org.lwjgl.opengl.GL11;
 
 import com.google.common.collect.Lists;
+import com.google.common.collect.Maps;
 
 import eu.gir.girsignals.EnumSignals.IIntegerable;
 import eu.gir.girsignals.GirsignalsMain;
@@ -136,6 +139,20 @@ public class GuiSignalController extends GuiContainer {
 		int index = 0;
 		pageList.clear();
 
+		final HashMap<SEProperty<?>, Object> map = Maps.newHashMap();
+		for(Entry<Integer, Integer> entry : sigController.guiCacheList) {
+			SEProperty<?> prop = SEProperty.cst(signal.getPropertyFromID(entry.getKey()));
+			map.put(prop, prop.getObjFromID(entry.getValue()));
+		}
+
+		for (int i = 0; i < properties.size(); i++) {
+			SEProperty<?> prop = SEProperty.cst(properties.get(i));
+			int sigState = sigController.supportedSigStates[i];
+			if (sigState < 0 || sigState >= prop.count())
+				continue;
+			map.put(prop, prop.getObjFromID(sigState));
+		}
+		
 		if (sigController.indexMode == EnumMode.MANUELL) {
 			pageList.add(Lists.newArrayList());
 			boolean visible = true;
@@ -144,7 +161,7 @@ public class GuiSignalController extends GuiContainer {
 				final int id = i;
 				final SEProperty<?> prop = SEProperty.cst(signal.getPropertyFromID(sigController.supportedSigTypes[i]));
 				final int state = sigController.supportedSigStates[i];
-				if (state < 0 || state >= prop.count())
+				if (state < 0 || state >= prop.count() || !prop.test(map))
 					continue;
 				if (!prop.isChangabelAtStage(ChangeableStage.APISTAGE) && !prop.isChangabelAtStage(ChangeableStage.APISTAGE_NONE_CONFIG))
 					continue;
@@ -402,6 +419,11 @@ public class GuiSignalController extends GuiContainer {
 		final Signal signal = Signal.SIGNALLIST.get(sigController.signalType);
 		IExtendedBlockState ebs = (IExtendedBlockState) signal.getDefaultState();
 
+		for(Entry<Integer, Integer> entry : sigController.guiCacheList) {
+			SEProperty prop = SEProperty.cst(signal.getPropertyFromID(entry.getKey()));
+			ebs = ebs.withProperty(prop, prop.getObjFromID(entry.getValue()));
+		}
+		
 		for (int i = 0; i < properties.size(); i++) {
 			SEProperty prop = SEProperty.cst(properties.get(i));
 			int sigState = sigController.supportedSigStates[i];